+++ app/vmui/packages/vmui/src/components/Main/Icons/index.tsx
@@ -113,6 +113,15 @@ export const ArrowDropDownIcon = () => (
   </svg>
 );
 
+export const ArrowBackIcon = () => (
+  <svg
+    viewBox="0 0 24 24"
+    fill="currentColor"
+  >
+    <path d="M20 11H7.83l5.59-5.59L12 4l-8 8 8 8 1.41-1.41L7.83 13H20z"></path>
+  </svg>
+);
+
 export const ClockIcon = () => (
   <svg
     viewBox="0 0 24 24"